Pipeline and services refactor 3: JD keyword extraction
