+++ requirements.txt
@@ -25,10 +25,6 @@ urllib3==2.0.7
 # System Monitoring
 psutil==5.9.5
 
-# Natural Language Processing
-nltk==3.8.1
-spacy==3.7.2
-
 # Database
 supabase==2.15.0
 
@@ -47,6 +43,5 @@ pandas==2.2.0
 pdflatex==0.1.3        # Python wrapper for pdflatex (optional, system pdflatex preferred)
 
 # Utility Libraries
-tqdm==4.66.1           # Progress bars
 packaging>=21.0        # Added for version comparison
